small fixes

+++ defenceSolution/defence/Enemy.cpp
@@ -44,7 +44,7 @@ void Enemy::teePolku(int pathnumber){
 
 void Enemy::draw(SDL_Renderer* rndr)
 {
-	SDL_Point p = {_location.w / 2, _location.h / 2};
+	SDL_Point p = {(int)_location.w / 2, (int)_location.h / 2};
 	SDL_Rect r = _location.toSDL_Rect();
 	SDL_RenderCopyEx(rndr, _texture, nullptr, &r, 0, &p, SDL_FLIP_NONE);
 }
